Refactor: Simplify program logic and remove unused certificates
