+++ package.json
@@ -2,10 +2,10 @@
   "name": "certmaker",
   "version": "1.2.0",
   "description": "Create Chingu Voyage Certificates",
-  "main": "src/index.js",
+  "main": "src/certmaker.js",
   "type": "module",
   "scripts": {
-    "start": "node src/index.js",
+    "start": "node src/certmaker.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "Jim D. Medlock",
